docs: auto-generate support h2 below all blog posts
