+++ packages/website/blog/2023-02-24-consistent-type-exports-and-imports-why-and-how.md
@@ -183,9 +183,3 @@ If it detects an import that only imports specifiers with inline `type` qualifie
 
 You can read more about the rules' configuration options in their docs pages.
 See [our Getting Started docs](/getting-started) for more information on linting your TypeScript code with typescript-eslint.
-
-## Supporting typescript-eslint
-
-If you enjoyed this blog post and/or use typescript-eslint, please consider [supporting us on Open Collective](https://opencollective.com/typescript-eslint).
-We're a small volunteer team and could use your support to make the ESLint experience on TypeScript great.
-Thanks! 💖
