+++ packages/website/blog/2024-05-27-announcing-typescript-eslint-v8-beta.mdx
@@ -349,9 +349,3 @@ We'd like to extend a sincere _thank you_ to everybody who pitched in to make ty
   <!-- cspell:enable -->
 
 See the [v8.0.0 milestone](https://github.com/typescript-eslint/typescript-eslint/milestone/9) for the list of issues and associated merged pull requests.
-
-## Supporting typescript-eslint
-
-If you enjoyed this blog post and/or use typescript-eslint, please consider [supporting us on Open Collective](https://opencollective.com/typescript-eslint).
-We're a small volunteer team and could use your support to make the ESLint experience on TypeScript great.
-Thanks! 💖
