+++ packages/website/plugins/generated-rule-docs/addESLintHashToCodeBlocksMeta.ts
@@ -1,8 +1,9 @@
 import type { MdxJsxFlowElement } from 'mdast-util-mdx';
 import type * as unist from 'unist';
 
+import { nodeIsCode } from '../utils/nodes';
+import { convertToPlaygroundHash } from '../utils/rules';
 import type { RuleDocsPage } from './RuleDocsPage';
-import { convertToPlaygroundHash, nodeIsCode } from './utils';
 
 const optionRegex = /option='(?<option>.*?)'/;
