+++ packages/website/plugins/generated-rule-docs/createRuleDocsPage.ts
@@ -1,10 +1,10 @@
 import type * as mdast from 'mdast';
 import type * as unist from 'unist';
 
+import type { RuleModuleWithMetaDocs, VFileWithStem } from '../utils/rules';
+import { findH2Index } from '../utils/rules';
 import type { HeadingName } from './RuleDocsPage';
 import { requiredHeadingNames, RuleDocsPage } from './RuleDocsPage';
-import type { RuleModuleWithMetaDocs, VFileWithStem } from './utils';
-import { findH2Index } from './utils';
 
 export function createRuleDocsPage(
   children: unist.Node[],
